+++ packages/devtools_shared/lib/src/deeplink/deeplink_manager.dart
@@ -51,7 +51,6 @@ class DeeplinkManager {
       };
     }
     return <String, String?>{
-      kErrorField: null,
       kOutputJsonField: outputJson,
     };
   }